requirements.txt
python-telegram-bot==21.6
pandas>=2.0.0
numpy>=1.24.0
